Misc int types rework

+++ bindings/src/utils/buffer_utils.c
@@ -22,10 +22,10 @@
 #include "buffer_utils.h"
 
 
-uint32_t import_from_pybuffer(const sipAPIDef* sipAPI, uint8_t **buf, PyObject* exporter)
+Py_ssize_t import_from_pybuffer(const sipAPIDef* sipAPI, uint8_t **buf, PyObject* exporter)
 {
     Py_buffer buffer;
-    uint32_t len;
+    Py_ssize_t len;
 
     if (*buf != NULL)
         sipAPI->api_free(*buf);
@@ -45,7 +45,7 @@ uint32_t import_from_pybuffer(const sipAPIDef* sipAPI, uint8_t **buf, PyObject* 
 }
 
 
-PyObject* export_to_pybuffer(const sipAPIDef* sipAPI, uint8_t *data, uint32_t len)
+PyObject* export_to_pybuffer(const sipAPIDef* sipAPI, uint8_t *data, Py_ssize_t len)
 {
     if (len > 0 && data) {
         void* buf = sipAPI->api_malloc(len);
